requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
redis==5.2.0
python-dotenv==1.0.1
jinja2==3.1.4
pydantic==2.10.0
python-multipart==0.0.12
httpx==0.27.0
